Extract MarketPriceAndSize Strategy
